feat: store вынесен в отдельный файл store.ts
