+++ src/components/app-header/app-header.tsx
@@ -1,12 +1,13 @@
+import { FC } from 'react';
 import { useSelector } from 'react-redux';
 import appHeaderStyle from  './app-header.module.css';
 import { NavLink } from 'react-router-dom';
 import { BurgerIcon, Logo, ListIcon, ProfileIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 import { PATH } from '../../utils/api';
 
-const AppHeader = () => {
-    const auth = useSelector((state) => state.auth.auth);
-    const name = useSelector((state) => state.auth.user.name);
+const AppHeader: FC = () => {
+    const auth = useSelector((state: any) => state.auth.auth);
+    const name = useSelector((state: any) => state.auth.user.name);
 
     const pendingLink = `${appHeaderStyle.header__item} ${appHeaderStyle.header__link} pb-4 pt-4 pl-5 pr-5`;
     const activeLink = `${appHeaderStyle.header__link_active} ${appHeaderStyle.header__item} ${appHeaderStyle.header__link} pb-4 pt-4 pl-5 pr-5`;
