+++ src/index.tsx
@@ -1,61 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { compose, legacy_createStore as createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 import './index.css';
 import { App } from './components';
 import reportWebVitals from './reportWebVitals';
-import { rootReducer } from './services/reducers/index';
 import { HashRouter } from 'react-router-dom';
-import { socketMiddleware } from './services/middleware/socketMiddleware';
-import {
-  WS_CONNECTION_START,
-  WS_CONNECTION_SUCCESS,
-  WS_CONNECTION_ERROR,
-  WS_CONNECTION_CLOSED,
-  WS_GET_MESSAGE,
-  WS_SEND_MESSAGE
-} from '../src/services/actions/wsActionTypes';
-
-import {
-  WS_AUTH_CONNECTION_START,
-  WS_AUTH_CONNECTION_SUCCESS,
-  WS_AUTH_CONNECTION_ERROR,
-  WS_AUTH_CONNECTION_CLOSED,
-  WS_AUTH_GET_MESSAGE,
-  WS_AUTH_SEND_MESSAGE
-} from './services/actions/wsAuthActionTypes';
-
-const wsActions = {
-  wsStart: WS_CONNECTION_START,
-  onOpen: WS_CONNECTION_SUCCESS,
-  onError: WS_CONNECTION_ERROR,
-  onClose: WS_CONNECTION_CLOSED,
-  onMessage: WS_GET_MESSAGE,
-  wsSend: WS_SEND_MESSAGE
-}
-
-const wsAuthActions = {
-  wsStart: WS_AUTH_CONNECTION_START,
-  onOpen: WS_AUTH_CONNECTION_SUCCESS,
-  onError: WS_AUTH_CONNECTION_ERROR,
-  onClose: WS_AUTH_CONNECTION_CLOSED,
-  onMessage: WS_AUTH_GET_MESSAGE,
-  wsSend: WS_AUTH_SEND_MESSAGE
-}
-
-declare global {
-  interface Window {
-    __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof compose;
-  }
-}
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-
-const enhancer = composeEnhancers(applyMiddleware(thunk, socketMiddleware(wsActions), socketMiddleware(wsAuthActions)));
-
-const store = createStore(rootReducer, enhancer);
+import { store } from './services/store';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root')  as HTMLElement
