+++ src/services/actions/wsAuthActionTypes.js
@@ -7,3 +7,12 @@ export const WS_AUTH_SEND_MESSAGE = 'WS_AUTH_SEND_MESSAGE';
 
 export const wsAuthConnectionStart = (url) => ({ type: WS_AUTH_CONNECTION_START, payload: url });
 export const wsAuthConnectionClosed = () => ({ type: WS_AUTH_CONNECTION_CLOSED });
+
+export const wsAuthActions = {
+    wsStart: WS_AUTH_CONNECTION_START,
+    onOpen: WS_AUTH_CONNECTION_SUCCESS,
+    onError: WS_AUTH_CONNECTION_ERROR,
+    onClose: WS_AUTH_CONNECTION_CLOSED,
+    onMessage: WS_AUTH_GET_MESSAGE,
+    wsSend: WS_AUTH_SEND_MESSAGE
+  }
